Add endpoint to get latest position history for asset

+++ Models/AssetPositionHistory.cs
@@ -7,9 +7,9 @@ public partial class AssetPositionHistory
 {
     public int Id { get; set; }
 
-    public int? X { get; set; }
+    public double? X { get; set; }
 
-    public int? Y { get; set; }
+    public double? Y { get; set; }
 
     public DateTime? DateTime { get; set; }
